+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Upgraded pdm and checkout actions to clean out nodejs deprecation warnings
 - Removed `logzero` dependency
 - Dropped Python 3.7 support as it has been EOL since June 2023
+- Read version from scm and package via `importlib_metadata`
 
 ### Fixed